Add endianness conversion, correctly sending and receiving packets, change types

+++ p1/icmp_send.h
@@ -2,6 +2,6 @@
 #define ICMP_SEND_H
 #include <netinet/ip_icmp.h>
 
-ssize_t send_packet (int sockfd, int seq, char* ip_addr);
+ssize_t send_packet (int sockfd, u_int16_t seq, char* ip_addr);
 
 #endif
